windows: fix multi-monitor wallpaper support
The monitor_id returned by IDesktopWallpaper is not the same as winit's
native_id, which is why this would fail to grab wallpapers from other
monitors.
The API is index-based, which is really annoying. Maybe
there's a way to get this ID from another API, but I haven't seen this
done.

Of note here is that the OsString approach in windows-wallpaper does not
work correctly: something in the encoding/decoding is going wrong. It
could be the trailing 0 byte, which is removed from PCWSTR when calling
as_wide, but reinserted in the hstring conversion.